문제 해결

+++ Holder/Holder/Common/Repositories/BookmarkRepository.swift
@@ -12,10 +12,11 @@ protocol BookmarkRepositoryType {
     func fetchedResultsController(for tag: Tag) -> NSFetchedResultsController<BookmarkTagEntity>
     func fetchedResultsController(for searchTerm: String, isFavorite: Bool) -> NSFetchedResultsController<BookmarkEntity>
     func isExisting(_ url: URL) -> Result<Bool, Error>
-    func add(with bookmark: Bookmark) -> Result<Void, Error>
-    func update(with bookmark: Bookmark) -> Result<Void, Error>
-    func update(_ bookmarkEntity: BookmarkEntity) -> Result<Bool, Error>
-    func delete(_ bookmarkEntity: BookmarkEntity) -> Result<Void, Error>
+    func add(bookmark: Bookmark) -> Result<Void, Error>
+    func update(bookmark: Bookmark) -> Result<Void, Error>
+    func updateFavorites(for url: URL) -> Result<Bool, Error>
+    func updateTags(_ tag: Tag, to newTag: Tag) -> Result<Void, Error>
+    func delete(for url: URL) -> Result<Void, Error>
     func deleteTags(for tag: Tag)
     func clear()
 }
@@ -24,9 +25,9 @@ final class BookmarkRepository: BookmarkRepositoryType {
     
     static let shared = BookmarkRepository()
     
-    var context: NSManagedObjectContext { PersistentContainer.shared.context }
+    private var context: NSManagedObjectContext { PersistentContainer.shared.context }
     
-    var ascending: Bool {
+    private var ascending: Bool {
         let sort = UserDefaults.value(forType: Sort.self) ?? .newestToOldest
         return sort == .newestToOldest ? false : true
     }
@@ -35,14 +36,14 @@ final class BookmarkRepository: BookmarkRepositoryType {
     
     func fetchedResultsController(isFavorite: Bool = false) -> NSFetchedResultsController<BookmarkEntity> {
         let request = BookmarkEntity.fetchRequest()
+        let sortDescriptor = NSSortDescriptor(key: #keyPath(BookmarkEntity.creationDate), ascending: ascending)
+        request.sortDescriptors = [sortDescriptor]
         
         if isFavorite {
             let predicate = NSPredicate(format: "%K == YES", #keyPath(BookmarkEntity.isFavorite))
             request.predicate = predicate
         }
         
-        let sortDescriptor = NSSortDescriptor(key: #keyPath(BookmarkEntity.creationDate), ascending: ascending)
-        request.sortDescriptors = [sortDescriptor]
         return NSFetchedResultsController(fetchRequest: request,
                                           managedObjectContext: context,
                                           sectionNameKeyPath: nil,
@@ -51,10 +52,10 @@ final class BookmarkRepository: BookmarkRepositoryType {
     
     func fetchedResultsController(for tag: Tag) -> NSFetchedResultsController<BookmarkTagEntity> {
         let request = BookmarkTagEntity.fetchRequest()
-        let predicate = NSPredicate(format: "%K == %@", #keyPath(BookmarkTagEntity.name), tag.name)
         let sortDescriptor = NSSortDescriptor(key: #keyPath(BookmarkTagEntity.bookmark.creationDate), ascending: ascending)
-        request.predicate = predicate
+        let predicate = NSPredicate(format: "%K == %@", #keyPath(BookmarkTagEntity.name), tag.name)
         request.sortDescriptors = [sortDescriptor]
+        request.predicate = predicate
         return NSFetchedResultsController(fetchRequest: request,
                                           managedObjectContext: context,
                                           sectionNameKeyPath: nil,
@@ -63,6 +64,7 @@ final class BookmarkRepository: BookmarkRepositoryType {
     
     func fetchedResultsController(for searchTerm: String, isFavorite: Bool = false) -> NSFetchedResultsController<BookmarkEntity> {
         let request = BookmarkEntity.fetchRequest()
+        let sortDescriptor = NSSortDescriptor(key: #keyPath(BookmarkEntity.creationDate), ascending: ascending)
         let titlePredicate = NSPredicate(format: "%K CONTAINS[cd] %@", #keyPath(BookmarkEntity.title), searchTerm)
         let urlPredicate = NSPredicate(format: "%K CONTAINS[cd] %@", #keyPath(BookmarkEntity.urlString), searchTerm)
         let notePredicate = NSPredicate(format: "%K CONTAINS[cd] %@", #keyPath(BookmarkEntity.note), searchTerm)
@@ -73,9 +75,9 @@ final class BookmarkRepository: BookmarkRepositoryType {
             predicates = NSCompoundPredicate(type: .and, subpredicates: [predicates, isFavoritePredicate])
         }
         
-        let sortDescriptor = NSSortDescriptor(key: #keyPath(BookmarkEntity.creationDate), ascending: ascending)
-        request.predicate = predicates
         request.sortDescriptors = [sortDescriptor]
+        request.predicate = predicates
+        
         return NSFetchedResultsController(fetchRequest: request,
                                           managedObjectContext: context,
                                           sectionNameKeyPath: nil,
@@ -94,7 +96,7 @@ final class BookmarkRepository: BookmarkRepositoryType {
         }
     }
     
-    func add(with bookmark: Bookmark) -> Result<Void, Error> {
+    func add(bookmark: Bookmark) -> Result<Void, Error> {
         let bookmarkEntity = BookmarkEntity(context: context)
         bookmarkEntity.configure(with: bookmark)
         do {
@@ -105,7 +107,7 @@ final class BookmarkRepository: BookmarkRepositoryType {
         }
     }
     
-    func update(with bookmark: Bookmark) -> Result<Void, Error> {
+    func update(bookmark: Bookmark) -> Result<Void, Error> {
         let request = BookmarkEntity.fetchRequest()
         let predicate = NSPredicate(format: "%K == %@", #keyPath(BookmarkEntity.urlString), bookmark.url.absoluteString)
         request.predicate = predicate
@@ -120,8 +122,12 @@ final class BookmarkRepository: BookmarkRepositoryType {
         }
     }
     
-    func update(_ bookmarkEntity: BookmarkEntity) -> Result<Bool, Error> {
+    func updateFavorites(for url: URL) -> Result<Bool, Error> {
+        let request = BookmarkEntity.fetchRequest()
+        let predicate = NSPredicate(format: "%K == %@", #keyPath(BookmarkEntity.urlString), url.absoluteString)
+        request.predicate = predicate
         do {
+            guard let bookmarkEntity = try context.fetch(request).first else { return .success(false) }
             bookmarkEntity.isFavorite.toggle()
             try context.save()
             return .success(bookmarkEntity.isFavorite)
@@ -130,8 +136,28 @@ final class BookmarkRepository: BookmarkRepositoryType {
         }
     }
     
-    func delete(_ bookmarkEntity: BookmarkEntity) -> Result<Void, Error> {
+    func updateTags(_ tag: Tag, to newTag: Tag) -> Result<Void, Error> {
+        let request = BookmarkTagEntity.fetchRequest()
+        let predicate = NSPredicate(format: "%K == %@", #keyPath(BookmarkTagEntity.name), tag.name)
+        request.predicate = predicate
         do {
+            let bookmarkTagEntities = try context.fetch(request)
+            if bookmarkTagEntities.count == 0 { return .success(()) }
+            bookmarkTagEntities.forEach { $0.name = newTag.name }
+            try context.save()
+            return .success(())
+        } catch {
+            return .failure(error)
+        }
+    }
+    
+    func delete(for url: URL) -> Result<Void, Error> {
+        let request = BookmarkEntity.fetchRequest()
+        let predicate = NSPredicate(format: "%K == %@", #keyPath(BookmarkEntity.urlString), url.absoluteString)
+        request.predicate = predicate
+        request.includesPropertyValues = false
+        do {
+            guard let bookmarkEntity = try context.fetch(request).first else { return .success(()) }
             context.delete(bookmarkEntity)
             try context.save()
             return .success(())
@@ -155,8 +181,13 @@ final class BookmarkRepository: BookmarkRepositoryType {
     }
     
     func clear() {
-        clearEntities(with: BookmarkEntity.fetchRequest())
-        clearEntities(with: BookmarkTagEntity.fetchRequest())
+        let fetchRequest: NSFetchRequest<NSFetchRequestResult> = BookmarkEntity.fetchRequest()
+        let deleteRequest = NSBatchDeleteRequest(fetchRequest: fetchRequest)
+        guard let _ = try? context.execute(deleteRequest) else {
+            NotificationCenter.post(named: NotificationName.didFailToProcessData)
+            return
+        }
+        NotificationCenter.post(named: NotificationName.Store.didSucceedToClear)
     }
     
     private func deleteTags(of bookmarkEntity: BookmarkEntity) throws {
@@ -167,13 +198,5 @@ final class BookmarkRepository: BookmarkRepositoryType {
         let bookmarkTagEntities = try context.fetch(request)
         bookmarkTagEntities.forEach { context.delete($0) }
         try context.save()
-    }
-    
-    private func clearEntities(with fetchRequest: NSFetchRequest<NSFetchRequestResult>) {
-        let deleteRequest = NSBatchDeleteRequest(fetchRequest: fetchRequest)
-        deleteRequest.resultType = .resultTypeObjectIDs
-        let deleteResult = try? context.execute(deleteRequest) as? NSBatchDeleteResult
-        guard let objectIDs = deleteResult?.result as? [NSManagedObjectID] else { return }
-        NSManagedObjectContext.mergeChanges(fromRemoteContextSave: [NSDeletedObjectsKey: objectIDs], into: [context])
     }
 }
